pyproject.toml
[project]
name = "polychrome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "sqlalchemy>=2.0.48",
    "pwdlib[argon2]>=0.3.0",
    "httpx>=0.28.1",
    "textual>=8.2.2",
    "argon2-cffi>=25.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-textual-snapshot>=1.0.0",
    "ruff>=0.15.9",
    "textual-dev",
]

[tool.pytest.ini_options]
testpaths = ["tests", "api/tests", "cli/tests"]
pythonpath = ["."]

[project.scripts]
polychrome-cli = "cli.app:main"

[tool.ruff]
line-length = 120
